requirements.txt
-e ../mezzanine
psycopg2==2.5.2
ipython==2.0.0
gunicorn==18.0
kitchen==1.1.1
vectorformats==0.1
simplejson==3.4.1
openpyxl==1.8.6
Fabric==1.8.3
django-modeltranslation==0.7.3
mezzanine-pagedown==0.7
-e git://github.com/pwhipp/mezzanine-slides.git#egg=mezzanine-slides
